data-validation-automation/requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.21.0

# Google Cloud
google-cloud-bigquery>=3.4.0
google-cloud-bigquery-storage>=2.13.0
google-auth>=2.6.0

# Data processing
pyarrow>=6.0.0
tqdm>=4.62.0

# Validation and analysis
jsonschema>=4.4.0
python-dateutil>=2.8.2

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development
black>=22.3.0
flake8>=4.0.0
mypy>=0.950

# Logging and monitoring
structlog>=21.5.0
rich>=12.0.0 